Fix batch processing level reporting in core configuration telemetry

+++ features/dd-sdk-android-rum/src/main/kotlin/com/datadog/android/telemetry/internal/TelemetryEventHandler.kt
@@ -263,7 +263,7 @@ internal class TelemetryEventHandler(
                     sessionReplaySampleRate = sessionReplaySampleRate,
                     defaultPrivacyLevel = sessionReplayPrivacy,
                     startSessionReplayRecordingManually = startSessionReplayManually,
-                    batchProcessingLevel = coreConfiguration.batchProcessingLevel
+                    batchProcessingLevel = coreConfiguration.batchProcessingLevel.toLong()
                 )
             )
         )
